# md/pauth-l2-openapi-v1.yaml
openapi: 3.1.0
info:
  title: pauth API
  version: 1.0.0
  description: |
    OpenAPI specification for the implemented pauth authentication model.

    - L1: gateway forward-auth (`GET /api/verify`)
    - OAuth: app login (`/api/l2/authorize`, `/api/l2/token`, `/api/l2/userinfo`)
    - Admin: clients, users, L1 permissions, invites (`/api/admin/*`)

    All active pauth users may OAuth into registered clients. Per-app authorization is managed by each application.

    OAuth `redirect_uri`: any valid HTTPS URL (localhost HTTP for dev). No per-client URI registration.
servers:
  - url: https://auth.example.com
    description: Production
...
paths:
...
  /api/admin/clients:
    get:
      tags: [Admin]
      summary: List OAuth clients
      security:
        - cookieSid: []
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      tags: [Admin]
      summary: Create OAuth client (auto-generates clientSecret)
      security:
        - cookieSid: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClientRequest"
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateClientResponse"
  /api/admin/clients/{clientId}:
    patch:
      tags: [Admin]
      summary: Update client
      security:
        - cookieSid: []
      parameters:
        - in: path
          name: clientId
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateClientRequest"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
    delete:
      tags: [Admin]
      summary: Delete client
      security:
        - cookieSid: []
      parameters:
        - in: path
          name: clientId
          required: true
          schema: { type: string }
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
  /api/admin/clients/{clientId}/regenerate-secret:
    post:
      tags: [Admin]
      summary: Regenerate client secret
      security:
        - cookieSid: []
      parameters:
        - in: path
          name: clientId
          required: true
          schema: { type: string }
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegenerateSecretResponse"
  /api/admin/users:
    get:
      tags: [Admin]
      summary: List users with L1 flag
      security:
        - cookieSid: []
      parameters:
        - in: query
          name: status
          schema: { type: string, enum: [pending, active, disabled, all] }
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminUser"
  /api/admin/users/{userId}:
    patch:
      tags: [Admin]
      summary: Update user name or status
      security:
        - cookieSid: []
      parameters:
        - in: path
          name: userId
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
    delete:
      tags: [Admin]
      summary: Delete user
      security:
        - cookieSid: []
      parameters:
        - in: path
          name: userId
          required: true
          schema: { type: string }
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
  /api/admin/users/{userId}/permissions:
    put:
      tags: [Admin]
      summary: Set user L1 gateway access
      security:
        - cookieSid: []
      parameters:
        - in: path
          name: userId
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetPermissionsRequest"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SetPermissionsResponse"
  /api/admin/invites:
    post:
      tags: [Admin]
      summary: Create pre-provisioned user and invite URL
      security:
        - cookieSid: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInviteRequest"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateInviteResponse"
components:
  securitySchemes:
    cookieSid:
      type: apiKey
      in: cookie
      name: sid
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  schemas:
...
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok: { type: boolean, const: true }
...
    AccessMode:
      type: string
      enum: [L2_ONLY, L1_AND_L2]
...
    Client:
      type: object
      required: [id, clientId, name, accessMode, enabled, redirectUris, clientSecret]
      properties:
        id: { type: string }
        clientId: { type: string }
        name: { type: string }
        accessMode:
          $ref: "#/components/schemas/AccessMode"
        enabled: { type: boolean }
        redirectUris:
          type: array
          items: { type: string }
          description: Legacy field, always empty array
        clientSecret:
          type: string
          nullable: true
          description: Plaintext secret for admin; null if not yet generated
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CreateClientRequest:
      type: object
      required: [clientId, name, accessMode]
      properties:
        clientId: { type: string }
        name: { type: string }
        accessMode:
          $ref: "#/components/schemas/AccessMode"
        enabled: { type: boolean, default: true }
    CreateClientResponse:
      type: object
      required: [ok, id, clientId, clientSecret]
      properties:
        ok: { type: boolean, const: true }
        id: { type: string }
        clientId: { type: string }
        clientSecret: { type: string }
    UpdateClientRequest:
      type: object
      properties:
        name: { type: string }
        accessMode:
          $ref: "#/components/schemas/AccessMode"
        enabled: { type: boolean }
    RegenerateSecretResponse:
      type: object
      required: [ok, clientId, clientSecret]
      properties:
        ok: { type: boolean, const: true }
        clientId: { type: string }
        clientSecret: { type: string }
    AdminUser:
      type: object
      properties:
        id: { type: string }
        email: { type: string }
        name: { type: string }
        role: { type: string }
        status: { type: string }
        passkeyCount: { type: integer }
        hasPendingInvite: { type: boolean }
        l1Enabled: { type: boolean }
    UpdateUserRequest:
      type: object
      properties:
        name: { type: string }
        status: { type: string, enum: [active, disabled] }
    SetPermissionsRequest:
      type: object
      required: [l1Enabled]
      properties:
        l1Enabled: { type: boolean }
    SetPermissionsResponse:
      type: object
      required: [ok, permissions]
      properties:
        ok: { type: boolean, const: true }
        permissions:
          type: object
          properties:
            l1Enabled: { type: boolean }
    CreateInviteRequest:
      type: object
      required: [name]
      properties:
        name: { type: string }
        role: { type: string, enum: [admin, user], default: user }
        l1Enabled: { type: boolean, default: false }
    CreateInviteResponse:
      type: object
      required: [ok, url, token, expiresAt, userId]
      properties:
        ok: { type: boolean, const: true }
        url: { type: string, format: uri }
        token: { type: string }
        expiresAt: { type: string, format: date-time }
        userId: { type: string }